backend/src/main/java/be/pxl/researchproject/api/request/ContactPersonRequest.java
package be.pxl.researchproject.api.request;

public class ContactPersonRequest {
    private String title, name, firstName, phoneNumber, email;
    private long Id;

    public ContactPersonRequest(String title, String name, String firstName, String phoneNumber, String email) {
        this.name = name;
        this.firstName = firstName;
        this.phoneNumber = phoneNumber;
        this.title = title;
        this.email = email;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public long getId() {
        return Id;
    }

    public void setId(long id) {
        Id = id;
    }
}
